chore: rename package to published name
